+++ react/components/BuyButton/mutations/setOpenMinicart.gql
@@ -1,0 +1,3 @@
+mutation setMinicartOpen($isOpen: Boolean!) {
+  setMinicartOpen(isOpen: $isOpen) @client
+}
